fix: adjust rpc packet define.
