+++ src/storage/cache/mod.rs
@@ -1,2 +1,5 @@
 /// RPC module for cache
 pub mod rpc;
+
+/// Unit tests for cache module
+pub mod tests;
