+++ src/storage/cache/tests/mod.rs
@@ -1,0 +1,2 @@
+/// Unit tests for cache module
+pub mod rpc;
